ml-api/requirements.txt

Flask==3.0.0
numpy
pandas
scikit-learn
xgboost
joblib
python-dotenv
